requirements.txt
langchain==0.1.17
langchain-core==0.1.48
langchain-community==0.0.36
langchain-openai==0.1.4
qdrant-client==1.9.0
tiktoken==0.6.0
pymupdf==1.24.2
python-dotenv==1.0.1
chainlit==0.7.700
openai==1.24.1
